Eliminate `walker` tests; create `refs` tests
